requirements.txt
aws-lambda-typing==2.20.0
boto3==1.42.5
boto3-stubs==1.42.3
botocore==1.42.5
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Deprecated==1.3.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==6.0.2
MarkupSafe==3.0.3
moto==5.1.18
packaging==25.0
pikepdf==10.0.2
pillow==12.0.0
pluggy==1.6.0
py-partiql-parser==0.6.3
pycparser==2.23
Pygments==2.19.2
pytest==9.0.1
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
responses==0.25.8
s3transfer==0.16.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.4
wrapt==2.0.1
xmltodict==1.0.2
